requirements.txt
flask==2.2.5
flask-cors==3.0.10
requests==2.31.0
python-dotenv==1.0.1
